/professor api 폴더구조 정리
