+++ apis/v2/professor/[id].ts
@@ -1,0 +1,15 @@
+import { getRequest } from '@/apis';
+import { FETCH_TAG_FACULTY } from '@/constants/network';
+import { WithLanguage } from '@/types/language';
+import { EmeritusFaculty, Faculty } from '@/types/people';
+
+// TODO: 왜 별도 함수로 있죠...?
+export const getFaculty = (id: number) =>
+  getRequest<WithLanguage<Faculty>>(`/v2/professor/${id}`, undefined, {
+    next: { tags: [FETCH_TAG_FACULTY] },
+  });
+
+export const getEmeritusFaculty = (id: number) =>
+  getRequest<WithLanguage<EmeritusFaculty>>(`/v2/professor/${id}`, undefined, {
+    next: { tags: [FETCH_TAG_FACULTY] },
+  });
