+++ apis/v2/professor/active.ts
@@ -1,0 +1,11 @@
+import { getRequest } from '@/apis';
+import { FETCH_TAG_FACULTY } from '@/constants/network';
+import { Language } from '@/types/language';
+import { FacultyList } from '@/types/people';
+
+export const getActiveFacultyList = (language: Language) =>
+  getRequest<FacultyList>(
+    '/v2/professor/active',
+    { language },
+    { next: { tags: [FETCH_TAG_FACULTY] } },
+  );
