+++ apis/v2/staff/[id].ts
@@ -1,0 +1,9 @@
+import { getRequest } from '@/apis';
+import { FETCH_TAG_STAFF } from '@/constants/network';
+import { WithLanguage } from '@/types/language';
+import { Staff } from '@/types/people';
+
+export const getStaff = (id: number) =>
+  getRequest<WithLanguage<Staff>>(`v2/staff/${id}`, undefined, {
+    next: { tags: [FETCH_TAG_STAFF] },
+  });
