+++ app/[locale]/people/faculty/page.tsx
@@ -1,4 +1,4 @@
-import { getActiveFacultyList } from '@/apis/people';
+import { getActiveFacultyList } from '@/apis/v2/professor/active';
 import { CreateButton } from '@/components/common/Buttons';
 import LoginVisible from '@/components/common/LoginVisible';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
